backend/ai-service/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
pydantic-settings

